bytte port bland annat. Pris kalkulering funkar nu

+++ webServiceClient/src/com/service/Itinerary/CheckAvailabilityResponse.java
@@ -17,7 +17,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="return" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="return" type="{http://www.w3.org/2001/XMLSchema}float"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -33,13 +33,13 @@ import javax.xml.bind.annotation.XmlType;
 public class CheckAvailabilityResponse {
 
     @XmlElement(name = "return")
-    protected int _return;
+    protected float _return;
 
     /**
      * Gets the value of the return property.
      * 
      */
-    public int getReturn() {
+    public float getReturn() {
         return _return;
     }
 
@@ -47,7 +47,7 @@ public class CheckAvailabilityResponse {
      * Sets the value of the return property.
      * 
      */
-    public void setReturn(int value) {
+    public void setReturn(float value) {
         this._return = value;
     }
